--- a/JASP-R-Interface/jaspResults/R/writeImage.R
+++ b/JASP-R-Interface/jaspResults/R/writeImage.R
@@ -263,10 +263,10 @@ decodeColNames <- function(x, strict = FALSE, fun = NULL, ...) {
     obj <- eval(parse(text = name))
   
   if (!exists(name) || !is.function(obj)) {
-    if (jaspResultsCalledFromJasp())
-      return(get(name, .GlobalEnv))
-    else
+    
+    if ("jasptools" %in% loadedNamespaces())
       return(getFromNamespace(name, asNamespace("jasptools")))
+    return(get(name, .GlobalEnv)) # works for both JASP and jaspTools
   }
   
   if (!is.function(obj))