--- a/JASP-Tests/R/tests/testthat/test-ancova.R
+++ b/JASP-Tests/R/tests/testthat/test-ancova.R
@@ -231,7 +231,7 @@ test_that("Analysis handles errors", {
  options$modelTerms <- list(list(components="contBinom"))
  results <- jasptools::run("Ancova", "test.csv", options)
  expect_identical(results[["results"]][["errorMessage"]], 
-                  "The following problem(s) occurred while running the analysis:<ul><li>Infinity found in debInf</li><li>Number of factor levels is < 2 in debInf</li></ul>",
+                  "The following problem(s) occurred while running the analysis:<ul><li>Infinity found in debInf</li></ul>",
                   label="Inf covariate check")
 
 })