--- a/JASP-Tests/R/tests/testthat/test-anova.R
+++ b/JASP-Tests/R/tests/testthat/test-anova.R
@@ -285,8 +285,8 @@ test_that("Analysis handles errors", {
   options$fixedFactors <- "contBinom"
   options$modelTerms <- list(list(components="contBinom"))
   results <- jasptools::run("Anova", "test.csv", options)
-  expect_identical(results[["results"]][["errorMessage"]], 
-                   "The following problem(s) occurred while running the analysis:<ul><li>Infinity found in debInf</li><li>Number of factor levels is < 2 in debInf</li></ul>",
+  expect_identical(results[["results"]][["errorMessage"]],
+                   "The following problem(s) occurred while running the analysis:<ul><li>Infinity found in debInf</li></ul>",
                    label="Inf dependent check")
 
   options$dependent <- "contNormal"
@@ -297,6 +297,7 @@ test_that("Analysis handles errors", {
   expect_identical(results[["results"]][["errorMessage"]], 
                    "The following problem(s) occurred while running the analysis:<ul><li>Infinity found in debInf</li></ul>",
                   label="Inf WLS weights check")
+  options$wlsWeights <- ""
 
   options$dependent <- "contNormal"
   options$fixedFactors <- "debSame"