Should fix https://github.com/jasp-stats/jasp-issues/issues/918 (but let's not close that issue when merging this, let's wait with that until the contributors in that thread can verify the timings).

Here a zip file with the profiling of this PR vs Stable.

[profileAnova.zip](https://github.com/jasp-stats/jasp-desktop/files/5090352/profileAnova.zip)

The slowdown/ speedup in a nutshell:

Take the flame graph for stable:

![image](https://user-images.githubusercontent.com/21319932/90512920-2ceb9200-e15f-11ea-9fd4-b21d5113e583.png)

The total time is 20930 ms of which 15340 ms are spend inside `.anovaCheckErrors`. If we zoom in on all the tiny bars we see:

![image](https://user-images.githubusercontent.com/21319932/90513211-8bb10b80-e15f-11ea-9cb4-aeb998747a46.png)

For some reason, `.v` is executed a million times (or more) and it's rather slow. The culprit here seems `.findFun` / `getAnywhere`. I've changed that into:

```r
if (jaspResultsCalledFromJasp())
  return(get(name, .GlobalEnv))
else
  return(getFromNamespace(name, asNamespace("jasptools")))
```

Furthermore, `.v` is called so often because `.applyOnGroups` calls it inside its lambda. I've now ensured that it's only called once outside of the lambda.

A final optimization I did was inside `.anovaCheckErrors`. The key there is that multiple calls to `.hasErrors` could be combined to one and that we don't need to loop over all model terms, but instead over the model terms with highest interactions. For example, if the variance is nonzero after grouping on A & B & C, it's certain that the variance is also nonzero when grouping on A, B, C, A & B, A & C, and B & C (which is what the old code did twice).


Here's the flame graph for this PR:

![image](https://user-images.githubusercontent.com/21319932/90513781-7688ac80-e160-11ea-8284-0b9538fc54cb.png)

The total run time is now 5390 ms of which 360 ms are spend inside `.anovaCheckErrors`.